refactor - Don't export `runLegionary` ...
... or, consequently, `RequestMsg`. Also, move `RequestMsg` from
`Network.Legion.Application` to `Network.Legion.Runtime`, because it is
strictly a runtime concern now that it is no longer part of the exposed
application interface.

The justification is that the use of this function (by a user) was highly
discouraged from the start, indicating it is probably a bad idea to
expose it. Also, maintaining two different entry points to the underling
framework is a bad idea.
